+++ NEWS.md
@@ -4,6 +4,7 @@
 AppVeyor.
 * Added precomp function to avoid repeated computation.
 * mrs_data2mat function now collapses all dimensions to dynamics by default.
+* mrs_data objects now store the nucleus.
 
 # spant 1.7.0
 * Added write_mrs function which guesses the output format from the file